trackers/ups/requirement.txt
certifi==2020.12.5
chardet==4.0.0
grpcio==1.36.1
grpcio-tools==1.36.1
idna==2.10
protobuf==3.15.6
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
urllib3==1.26.4
